+++ packages/gatsby-theme-carbon/index.js
@@ -16,5 +16,7 @@ export { default as ImageCard } from './src/components/ImageCard';
 export { default as SquareCard } from './src/components/SquareCard';
 export { Tabs, Tab } from './src/components/Tabs';
 export { default as InlineNotification } from './src/components/InlineNotification';
+export { default as ExpressiveListContainer } from './src/components/ExpressiveListContainer';
+export { default as ExpressiveList } from './src/components/ExpressiveList';
 // Homepage Template Components
 export { HomepageCallout, HomepageBanner } from './src/components/Homepage';