+++ packages/gatsby-theme-carbon/src/components/MDXProvider/defaultComponents.js
@@ -31,6 +31,8 @@ import { MiniCard, CardGroup } from '../MiniCard';
 import Sup from '../markdown/Sup';
 import Abbr from '../markdown/Abbr';
 import SquareCard from '../SquareCard';
+import ExpressiveListContainer from '../ExpressiveListContainer';
+import ExpressiveList from '../ExpressiveList';
 
 const components = {
   wrapper: function Wrapper({ children, ...props }) {
@@ -83,6 +85,8 @@ const components = {
   MiniCard,
   CardGroup,
   SquareCard,
+  ExpressiveListContainer,
+  ExpressiveList,
 };
 
 export default components;