+++ packages/gatsby-theme-carbon/src/components/Switcher/Switcher.js
@@ -96,6 +96,9 @@ const DefaultChildren = () => (
     <SwitcherLink href="https://www.carbondesignsystem.com/">
       Carbon Design System
     </SwitcherLink>
+    <SwitcherLink href="http://ibm.biz/carbon4ibmproducts" isInternal>
+      Carbon for IBM Products
+    </SwitcherLink>
     <SwitcherLink href="https://www.ibm.com/standards/carbon/">
       Carbon for IBM.com
     </SwitcherLink>
@@ -117,9 +120,6 @@ const DefaultChildren = () => (
     </SwitcherLink>
     <SwitcherLink href="https://www.ibm.com/design/research/">
       IBM Design Research
-    </SwitcherLink>
-    <SwitcherLink href="https://www.ibm.com/services/ibmix/">
-      IBM iX
     </SwitcherLink>
     <SwitcherLink
       isInternal